+++ Tests/TestSocketIO/SocketMangerTest.swift
@@ -167,6 +167,43 @@ class SocketMangerTest : XCTestCase {
         XCTAssertNil(manager.nsps[socket.nsp])
     }
 
+    func testAutoConnectFalseByDefault() {
+        let manager = SocketManager(socketURL: URL(string: "http://localhost")!, config: [])
+        XCTAssertFalse(manager.autoConnect)
+        XCTAssertEqual(manager.status, .notConnected, "manager should not auto-connect by default")
+        XCTAssertNil(manager.engine, "default config must NOT have created an engine")
+    }
+
+    func testAutoConnectExplicitFalse() {
+        let manager = SocketManager(
+            socketURL: URL(string: "http://localhost")!,
+            config: [.autoConnect(false)]
+        )
+        XCTAssertFalse(manager.autoConnect)
+        XCTAssertEqual(manager.status, .notConnected)
+    }
+
+    func testAutoConnectTrueTriggersConnect() {
+        let manager = SocketManager(
+            socketURL: URL(string: "http://localhost")!,
+            config: [.autoConnect(true)]
+        )
+        XCTAssertTrue(manager.autoConnect)
+        XCTAssertEqual(manager.status, .connecting,
+                       "autoConnect=true should put manager into .connecting immediately after init")
+        XCTAssertNotNil(manager.engine, "autoConnect=true should have triggered addEngine()")
+    }
+
+    func testAutoConnectFalseExplicitDoesNotTrigger() {
+        let manager = SocketManager(
+            socketURL: URL(string: "http://localhost")!,
+            config: [.autoConnect(false), .forceNew(true)]
+        )
+        XCTAssertEqual(manager.status, .notConnected)
+        XCTAssertTrue(manager.forceNew, "forceNew should still be honored independently")
+        XCTAssertNil(manager.engine, "autoConnect=false must NOT have created an engine")
+    }
+
     func testConnectSocketUsesExplicitPayloadWithRecoveryState() throws {
         let engine = CaptureEngine()
         manager.engine = engine